nyan: Clock setup.
I cleaned up the clk_rst struct: no need to have an array for the
source select because each thing is an individual entity.

There is a bit of a trick in here. I'm using macros to index arrays in the
clk_rst struct. Simple reason: if people use the constants we get compile-time
checking for simple errors.

init_pllx is fixed. The vendor table in u-boot was wrong.

BUG=None
TEST=builds. Amazing. Prints. Amazing.
BRANCH=None

Change-Id: I4429b3e5c14c200db6b4dd129eff14261fe1f326
Signed-off-by: Ronald G. Minnich <rminnich@google.com>
Reviewed-on: https://chromium-review.googlesource.com/172106
Tested-by: Gabe Black <gabeblack@chromium.org>
Reviewed-by: Ronald Minnich <rminnich@chromium.org>
Reviewed-by: Julius Werner <jwerner@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>